FCQ and Prof db layout complete
Created structure of tables for the fcqs and professors. Also reformated the text files used to populate the tables so the information would be filled correctly.
