+++ fcq/serializers.py
@@ -1,23 +1,9 @@
 from rest_framework import serializers
 
 
-class TeacherSerializer(serializers.Serializer):
+class ProfessorSerializer(serializers.Serializer):
     firstName = serializers.CharField()
     lastName = serializers.CharField()
-    mainDepartment = serializers.CharField()
-    numClasses = serializers.IntegerField() 
-    avgClassSize = serializers.IntegerField()
-    avgInstRating = serializers.FloatField()
-    avgCourseRating = serializers.FloatField()
-    avgChallenge = serializers.FloatField()
-    courseList = serializers.ListField(child=serializers.CharField())
-    timesCourseTaught = serializers.ListField(child=serializers.IntegerField())
-    courseRating = serializers.ListField(child=serializers.FloatField())
-    courseInstRating = serializers.ListField(child=serializers.FloatField())
-    courseChallenge = serializers.ListField(child=serializers.FloatField())
-    classIndex = serializers.ListField(child=serializers.IntegerField())
-
-
 
 class FcqSerializer(serializers.Serializer):
     index = serializers.IntegerField()
