+++ fcq/urls.py
@@ -7,6 +7,6 @@ app_name = "fcq"
 urlpatterns = [
     path("fcq_search/", views.fcq_search, name="fcq_search"),
     path("fcq_search_ajax/", views.fcq_search_ajax, name="fcq_search_ajax"),
-    path("view/<pk>/", views.TeacherView.as_view(), name="view"),
+    path("view/<pk>/", views.ProfessorView.as_view(), name="view"),
 ]
